+++ DbDwh/IIS/Functions/fnGetNomenclature.sql
@@ -1,37 +1,39 @@
-﻿-- [IIS].[fnGetNomenclatureByCode]
+﻿-- [IIS].[fnGetNomenclature]
 
 -- DROP FUNCTION
-DROP FUNCTION IF EXISTS [IIS].[fnGetNomenclatureByCode]
+DROP FUNCTION IF EXISTS [IIS].[fnGetNomenclature]
 GO
 
 -- CREATE FUNCTION
-CREATE FUNCTION [IIS].[fnGetNomenclatureByCode] (@code NVARCHAR(255)) RETURNS XML
-AS BEGIN
+CREATE FUNCTION [IIS].[fnGetNomenclature]
+(@code NVARCHAR(255) = NULL, @id BIGINT = NULL)
+RETURNS XML
+BEGIN
 	-- DECLARE.
 	-- Переменная с таблицей типов номенклатур
 	DECLARE @tableNomenclatureTypes TABLE 
 		([Name] VARCHAR(150), [GoodsForSale] BIT, [ID] INT, [CreateDate] DATETIME, [DLM] DATETIME, [StatusID] INT, [InformationSystemID] INT, [CodeInIS] VARBINARY(16))
 	INSERT INTO @tableNomenclatureTypes
 		SELECT [Name], [GoodsForSale], [ID], [CreateDate], [DLM], [StatusID], [InformationSystemID], [CodeInIS]
-		FROM [DW].[DimTypesOfNomenclature] READUNCOMMITTED 
+		FROM [DW].[DimTypesOfNomenclature] READUNCOMMITTED;
 	-- Переменная с таблицей плановой себестоимости
 	DECLARE @tableCurrentPlannedCosts TABLE 
 		([Marked] BIT, [Posted] BIT, [DocNum] VARCHAR(15), [DocDate] DATETIME, [Price] DECIMAL(15,3), [NomenclatureID] INT, [NomenclatureName] VARCHAR(150), [ID] BIGINT)
 	INSERT INTO @tableCurrentPlannedCosts
 		SELECT [Marked], [Posted], [DocNum], [DocDate], [Price], [NomenclatureID], [NomenclatureName], [ID]
-		FROM [DW].[vwCurrentPlannedCost] READUNCOMMITTED 
+		FROM [DW].[vwCurrentPlannedCost] READUNCOMMITTED;
 	-- Переменная с таблицей групп номенклатур
 	DECLARE @tableNomenclatureGroups TABLE 
 		([Name] NVARCHAR(150), [ID] INT, [CreateDate] DATETIME, [DLM] DATETIME, [StatusID] INT, [InformationSystemID] INT, [CodeInIS] VARBINARY(16))
 	INSERT INTO @tableNomenclatureGroups
 		SELECT [Name], [ID], [CreateDate], [DLM], [StatusID], [InformationSystemID],[CodeInIS]
-		FROM [DW].[DimNomenclatureGroups] READUNCOMMITTED 
+		FROM [DW].[DimNomenclatureGroups] READUNCOMMITTED;
 	-- Переменная с таблицей брендов
 	DECLARE @tableBrands TABLE 
 		([Code] NVARCHAR(15), [Name] NVARCHAR(150), [ID] INT, [CreateDate] DATETIME, [DLM] DATETIME, [StatusID] INT, [InformationSystemID] INT, [CodeInIS] VARBINARY(16))
 	INSERT INTO @tableBrands
 		SELECT [Code], [Name], [ID], [CreateDate], [DLM], [StatusID], [InformationSystemID], [CodeInIS]
-		FROM [DW].[DimBrands] READUNCOMMITTED
+		FROM [DW].[DimBrands] READUNCOMMITTED;
 	-- Переменная с таблицей себестоимости
 	DECLARE @tableSalfeCosts TABLE 
 		([PriceType] VARCHAR(5), [DateID] INT, [NomenclatureID] INT, [Nomenclature] NVARCHAR(150), [Price] DECIMAL(15,2)
@@ -39,7 +41,9 @@ AS BEGIN
 	INSERT INTO @tableSalfeCosts
 		SELECT [PriceType], [DateID], [NomenclatureID], [Nomenclature], [Price], [StartDate], [DLM]
 		FROM [VSDWH].[DW].[vwSelfCosts] READUNCOMMITTED
-		ORDER BY [PriceType] ASC
+		ORDER BY [PriceType] ASC;
+	-- Переменная с таблицей id
+	DECLARE @table_id TABLE ([ID] INT)
 	-- Переменная с таблицей фактических прайсов
 	DECLARE @tableFactPrices TABLE 
 		([DateID] INT, [DocNum] NVARCHAR(15), [DocDate] DATETIME, [DocType] NVARCHAR(100), [Marked] BIT, [Posted] BIT, [NomenclatureID] VARBINARY(16)
@@ -51,14 +55,14 @@ AS BEGIN
 		SELECT [DateID], [DocNum], [DocDate], [DocType], [Marked], [Posted], [NomenclatureID], [PriceTypeID], [DeliveryPlaceID], [_DateID], [_NomenclatureID], [_PriceTypeID], [_ContragentID]
 			, [_DeliveryPlaceID], [Price], [IsAction], [StartDate], [EndDate], [ID], [CreateDate], [DLM], [StatusID], [InformationSystemID], [CodeInIS], [_LineNo], [CHECKSUMM], [Comment]
 		FROM [DW].[FactPrices] READUNCOMMITTED 
-		WHERE [PriceTypeID]=0xBA6D90E6BA17BDD711E297052E5C534D -- Оптовые
-			AND [DocType]='DocumentRef.УстановкаЦенНоменклатуры' AND [IsAction]=0
-			AND [Marked]=0 AND [Posted]=1
+		WHERE [PriceTypeID] = 0xBA6D90E6BA17BDD711E297052E5C534D -- Оптовые
+			AND [DocType] = 'DocumentRef.УстановкаЦенНоменклатуры' AND [IsAction] = 0
+			AND [Marked] = 0 AND [Posted] = 1;
 	-- Переменная XML
 	DECLARE @xml XML = (
 		SELECT * FROM (
 			SELECT
-			 [N].[ID] "@ID"
+			 [N].[ID] "@id"
 			,[N].[Name] "@Name"
 			,[N].[Code] "@Code"
 			,[N].[MasterId] "@MasterId"
@@ -79,7 +83,7 @@ AS BEGIN
 			,CAST((SELECT * FROM ((
 				SELECT [PriceType] [@PriceType], [Price] [@Price], [StartDate] [@StartDate], [DLM] [@DLM]
 				FROM @tableSalfeCosts
-				WHERE [N].[ID]=[NomenclatureId]
+				WHERE [N].[ID] = [NomenclatureID]
 			)) [SelfCost] FOR XML PATH ('SelfCost'), BINARY BASE64) AS XML) [SelfCosts]
 			-- Раздел <Prices>
 			,CAST((SELECT
@@ -87,46 +91,55 @@ AS BEGIN
 					,[IsAction] [@IsAction]
 					,[StartDate] [@StartDate]
 				FROM @tableFactPrices [FP]
-				WHERE [FP].[PriceTypeID]=0xBA6D90E6BA17BDD711E297052E5C534D -- Оптовые
-					AND [FP].[DocType]='DocumentRef.УстановкаЦенНоменклатуры' AND [FP].[IsAction]=0
-					AND [FP].[Marked]=0 AND [FP].[Posted]=1
-					AND [N].[CodeInIS]=[FP].[NomenclatureID]
+				WHERE [FP].[PriceTypeID] = 0xBA6D90E6BA17BDD711E297052E5C534D -- Оптовые
+					AND [FP].[DocType] = 'DocumentRef.УстановкаЦенНоменклатуры' AND [FP].[IsAction] = 0
+					AND [FP].[Marked] = 0 AND [FP].[Posted] = 1
+					AND [N].[CodeInIS] = [FP].[NomenclatureID]
 				ORDER BY [StartDate]
 				FOR XML PATH ('Price'), BINARY BASE64)
 			AS XML) [Prices]
-		from [DW].[DimNomenclatures] [N]
-		LEFT JOIN @tableNomenclatureTypes [TN] ON [N].NomenclatureType = [TN].[CodeInIS] --AND Nomenclature.[InformationSystemID] = t.[InformationSystemID]
-		LEFT JOIN @tableCurrentPlannedCosts [vCPC] ON [N].[ID]=[vCPC].[NomenclatureID]
-		LEFT JOIN [DW].[FactPlannedCost] [FPC] ON [FPC].[ID]=[vCPC].[ID]
-		LEFT JOIN @tableNomenclatureGroups [NG] ON [N].[NomenclatureGroup]=[NG].[CodeInIS]
-		LEFT JOIN @tableBrands [B] ON [N].[Brand]=[B].[CodeInIS]
-		where 
+		FROM [DW].[DimNomenclatures] [N]
+		LEFT JOIN @tableNomenclatureTypes [NT] ON [N].NomenclatureType = [NT].[CodeInIS] --AND Nomenclature.[InformationSystemID] = t.[InformationSystemID]
+		LEFT JOIN @tableCurrentPlannedCosts [vCPC] ON [N].[ID] = [vCPC].[NomenclatureID]
+		LEFT JOIN @tableNomenclatureGroups [NG] ON [N].[NomenclatureGroup] = [NG].[CodeInIS]
+		LEFT JOIN @tableBrands [B] ON [N].[Brand] = [B].[CodeInIS]
+		WHERE
 			--JSON_VALUE([n].[Parents], '$.parents[0]') IN ('Колбасные изделия','Мясные продукты','Рыбная продукция')
-			[TN].[GoodsForSale]=1
-			AND COALESCE([N].[Marked], 0)=0
-			AND [N].[Code]=@code
-	) [DATA] FOR XML PATH ('Nomenclature'), ROOT('Goods'), BINARY BASE64)
+			[NT].[GoodsForSale] = 1
+			AND COALESCE([N].[Marked], 0) = 0
+			AND (  -- SWITCH @id OR @code
+				(@id IS NOT NULL AND [N].[ID] = @id) OR 
+				(@code IS NOT NULL AND [N].[Code] = @code)
+			)
+	) [DATA] FOR XML PATH ('Nomenclature'), ROOT('Goods'), BINARY BASE64);
 	-- RESULT.
-	DECLARE @Version NVARCHAR(100) = 'v.0.6.160'
-	DECLARE @Api NVARCHAR(1000) = '/api/nomenclature/?code=' + @code
+	DECLARE @version NVARCHAR(100) = 'v.0.6.170';
+	DECLARE @api NVARCHAR(1000)
+	IF (@code IS NOT NULL) BEGIN
+		SET @api = '/api/nomenclature/?code=' + @code;
+	END ELSE IF (@id IS NOT NULL) BEGIN
+		SET @api = '/api/nomenclature/?id=' + CAST(@id AS NVARCHAR(100));
+	END
 	IF (@xml IS NULL) BEGIN
-		SET @xml = (SELECT '' FOR XML PATH(''), ROOT('Response'), BINARY BASE64)
-		SET @xml.modify ('insert <Information /> as first into (./Response)[1] ')
-		SET @xml.modify ('insert attribute Version{sql:variable("@Version")} as first into (./Response/Information)[1] ')
-		SET @xml.modify ('insert (attribute Api{sql:variable("@Api")}, attribute ResultCount{0}) as last into (./Response/Information)[1] ')
+		SET @xml = (SELECT '' FOR XML PATH(''), ROOT('Response'), BINARY BASE64);
+		SET @xml.modify ('insert <Information /> as first into (./Response)[1] ');
+		SET @xml.modify ('insert attribute Version{sql:variable("@Version")} as first into (./Response/Information)[1] ');
+		SET @xml.modify ('insert (attribute Api{sql:variable("@Api")}, attribute ResultCount{0}) as last into (./Response/Information)[1] ');
 	END ELSE BEGIN
-		SET @xml.modify ('insert <Information /> as first into (./Goods)[1] ')
-		SET @xml.modify ('insert attribute Version{sql:variable("@Version")} as first into (./Goods/Information)[1] ')
-		SET @xml.modify ('insert (attribute Api{sql:variable("@Api")}, attribute ResultCount{count(.//Nomenclature)}) as last into (./Goods/Information)[1] ')
+		SET @xml.modify ('insert <Information /> as first into (./Goods)[1] ');
+		SET @xml.modify ('insert attribute Version{sql:variable("@Version")} as first into (./Goods/Information)[1] ');
+		SET @xml.modify ('insert (attribute Api{sql:variable("@Api")}, attribute ResultCount{count(.//Nomenclature)}) as last into (./Goods/Information)[1] ');
 	END
 	RETURN @xml
 END
 GO
 
 -- ACCESS
-GRANT EXECUTE ON [IIS].[fnGetNomenclatureByCode] TO [TerraSoftRole]
+GRANT EXECUTE ON [IIS].[fnGetNomenclature] TO [TerraSoftRole]
 GO
 
 -- CHECK FUNCTION
-DECLARE @Code NVARCHAR(255) = N'ЦБД00018851'
-SELECT [IIS].[fnGetNomenclatureByCode](@Code) [fnGetNomenclatureByCode]
+DECLARE @id BIGINT = -2147460739
+SELECT [IIS].[fnGetNomenclature] (DEFAULT, @id) [fnGetNomenclature]
+DECLARE @code NVARCHAR(255) = N'ЦБД00018851'
+SELECT [IIS].[fnGetNomenclature] (@code, DEFAULT) [fnGetNomenclature]