+++ WebApiTerra1000/Controllers/DeliveryPlaceControllerV2.cs
@@ -16,11 +16,11 @@ using static DataCore.ShareEnums;
 
 namespace WebApiTerra1000.Controllers
 {
-    public class DeliveryPlaceController : BaseController
+    public class DeliveryPlaceControllerV2 : BaseController
     {
         #region Constructor and destructor
 
-        public DeliveryPlaceController(ILogger<DeliveryPlaceController> logger, ISessionFactory sessionFactory) : base(logger, sessionFactory)
+        public DeliveryPlaceControllerV2(ILogger<DeliveryPlaceControllerV2> logger, ISessionFactory sessionFactory) : base(logger, sessionFactory)
         {
             //
         }
@@ -31,13 +31,24 @@ namespace WebApiTerra1000.Controllers
 
         [AllowAnonymous]
         [HttpGet()]
-        [Route("api/deliveryplaces/")]
+        [Route("api/v2/deliveryplaces/")]
         public ContentResult GetDeliveryPlaces(DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 100,
+            FormatType format = FormatType.Xml) =>
+            GetDeliveryPlacesWork(SqlQueriesV2.GetDeliveryPlaces, startDate, endDate, offset, rowCount, format);
+
+        [AllowAnonymous]
+        [HttpGet()]
+        [Route("api/v2/deliveryplaces_preview/")]
+        public ContentResult GetDeliveryPlacesPreview(DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 100,
+            FormatType format = FormatType.Xml) =>
+            GetDeliveryPlacesWork(SqlQueriesV2.GetDeliveryPlacesPreview, startDate, endDate, offset, rowCount, format);
+
+        private ContentResult GetDeliveryPlacesWork(string url, DateTime startDate, DateTime endDate, int offset = 0, int rowCount = 100,
             FormatType format = FormatType.Xml)
         {
             return Controller.RunTask(new Task<ContentResult>(() =>
             {
-                string response = TerraUtils.Sql.GetResponse<string>(SessionFactory, SqlQueries.GetDeliveryPlaces,
+                string response = TerraUtils.Sql.GetResponse<string>(SessionFactory, url,
                     TerraUtils.Sql.GetParameters(startDate, endDate, offset, rowCount));
                 XDocument xml = XDocument.Parse(response ?? $"<{TerraConsts.DeliveryPlaces} />", LoadOptions.None);
                 XDocument doc = new(new XElement(TerraConsts.Response, xml.Root));