+++ WebApiTerra1000/Utils/ControllerHelper.cs
@@ -60,7 +60,7 @@ namespace WebApiTerra1000.Common
                 GC.Collect();
             }
         }
-
+        
         #endregion
 
     }